requirements.txt
# API Clients
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.4.0
together>=0.2.0
aiohttp>=3.9.0

# Async
asyncio-throttle>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Statistics
scipy>=1.11.0
pingouin>=0.5.3
krippendorff>=0.6.0
statsmodels>=0.14.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
jsonschema>=4.17.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
